fix: :adhesive_bandage: Move request to service
